feat: refactor example ui
