+++ examples/react-ui/src/models/index.ts
@@ -1,1 +1,5 @@
 export * from './transfer';
+
+export type Props = {
+  readonly children: React.ReactNode;
+};
