+++ examples/react-ui/src/utils/solana.ts
@@ -10,10 +10,8 @@ import {
 } from '@solana/web3.js';
 import { SPLToken } from '@neonevm/token-transfer-core';
 import { sendSolanaTransaction } from './transfer';
-import { BaseMessageSignerWalletAdapter } from '@solana/wallet-adapter-base';
 
-export async function getOrCreateAssociatedTokenAccount(connection: Connection, signer: BaseMessageSignerWalletAdapter, token: SPLToken): Promise<AccountInfo<Buffer>> {
-  const solanaWallet = signer.publicKey!;
+export async function getOrCreateAssociatedTokenAccount(connection: Connection, signMethod: any, solanaWallet: PublicKey, token: SPLToken): Promise<AccountInfo<Buffer>> {
   const tokenMint = new PublicKey(token.address_spl);
   const tokenAccount = getAssociatedTokenAddressSync(tokenMint, solanaWallet);
   let account = await connection.getAccountInfo(tokenAccount);
@@ -21,7 +19,7 @@ export async function getOrCreateAssociatedTokenAccount(connection: Connection, 
     const transaction = new Transaction();
     transaction.add(createAssociatedTokenAccountInstruction(solanaWallet, tokenAccount, solanaWallet, tokenMint));
     transaction.recentBlockhash = (await connection.getLatestBlockhash('finalized')).blockhash;
-    const signature = await sendSolanaTransaction(connection, transaction, signer, true);
+    const signature = await sendSolanaTransaction(connection, transaction, signMethod, solanaWallet, true);
     account = await connection.getAccountInfo(tokenAccount);
     console.log(`Token Account created`, signature);
   }
