+++ examples/react-ui/src/utils/web3.ts
@@ -1,8 +1,9 @@
 import { erc20Abi, NEON_TOKEN_MINT_DECIMALS, SPLToken } from '@neonevm/token-transfer-core';
-import { Connection, LAMPORTS_PER_SOL, PublicKey, TokenAmount } from '@solana/web3.js';
+import { Connection, LAMPORTS_PER_SOL, PublicKey } from '@solana/web3.js';
 import { getAssociatedTokenAddress } from '@solana/spl-token';
 import { Contract, JsonRpcProvider, Wallet } from 'ethers';
 import { Big } from 'big.js';
+import { BIG_ZERO } from './consts.ts';
 
 export async function neonBalanceEthers(provider: JsonRpcProvider, address: Wallet | string): Promise<Big> {
   const balance = await provider.getBalance(address);
@@ -14,11 +15,16 @@ export async function solanaBalance(connection: Connection, address: PublicKey):
   return new Big(balance).div(LAMPORTS_PER_SOL);
 }
 
-export async function splTokenBalance(connection: Connection, walletPubkey: PublicKey, token: SPLToken): Promise<TokenAmount> {
-  const mintAccount = new PublicKey(token.address_spl);
-  const assocTokenAccountAddress = await getAssociatedTokenAddress(mintAccount, walletPubkey);
-  const response = await connection.getTokenAccountBalance(assocTokenAccountAddress);
-  return response?.value;
+export async function splTokenBalance(connection: Connection, walletPubkey: PublicKey, token: SPLToken): Promise<Big> {
+  try {
+    const mintAccount = new PublicKey(token.address_spl);
+    const assocTokenAccountAddress = await getAssociatedTokenAddress(mintAccount, walletPubkey);
+    const response = await connection.getTokenAccountBalance(assocTokenAccountAddress);
+    return new Big(response?.value.amount).div(Math.pow(10, response?.value.decimals));
+  } catch (e) {
+    console.error(`It seems that you haven\t ${token.symbol} tokens'${e}`);
+    return BIG_ZERO;
+  }
 }
 
 export async function mintTokenBalanceEthers(wallet: string, token: SPLToken, provider: JsonRpcProvider, contractAbi: any = erc20Abi, method = 'balanceOf'): Promise<Big> {
